pyproject.toml
[project]
name = "netfury"
version = "0.1.0"
description = "Benchmark 360 - Pipeline de extraccion automatizada de planes ISP competidores"
requires-python = ">=3.12"
dependencies = [
    # HTTP & scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "playwright>=1.42.0",
    "lxml>=4.9.0",
    # Data
    "pandas>=2.1.0",
    "pyarrow>=15.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # OCR local
    "pytesseract>=0.3.10",
    "easyocr>=1.7.0",
    "Pillow>=10.0.0",
    # LLM clients
    "anthropic>=0.45.0",
    "openai>=1.30.0",
    "google-genai>=1.0.0",
    # Web
    "flask>=3.0.0",
    # Config
    "python-dotenv>=1.0.0",
    # Notebook & viz
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    # Testing
    "pytest>=8.0.0",
]
